patch for bug 58840
git-svn-id: svn+ssh://svn.gentoo.org/var/svnroot/catalyst/trunk@423 d1e1f19c-881f-0410-ab34-b69fee027534

+++ targets/stage1/stage1-preclean2-chroot.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 # Copyright 1999-2004 Gentoo Technologies, Inc.
 # Distributed under the terms of the GNU General Public License v2
-# $Header: /var/cvsroot/gentoo/src/catalyst/targets/stage1/Attic/stage1-preclean2-chroot.sh,v 1.2 2004/05/17 01:21:17 zhen Exp $
+# $Header: /var/cvsroot/gentoo/src/catalyst/targets/stage1/Attic/stage1-preclean2-chroot.sh,v 1.3 2004/08/02 23:23:34 zhen Exp $
 		
 #now, some finishing touches to initialize gcc-config....
 unset ROOT
@@ -14,5 +14,9 @@ then
 fi
 
 #stage1 is not going to have anything in zoneinfo, so save our Factory timezone
-rm -f /etc/localtime
-cp /usr/share/zoneinfo/Factory /etc/localtime
+if [ -d /usr/share/zoneinfo ] ; then
+	rm -f /etc/localtime
+	cp /usr/share/zoneinfo/Factory /etc/localtime
+else
+	echo UTC > /etc/TZ
+fi
